Run all 9 scrapers in parallel + cut delays to 0.5s
- All scrapers run simultaneously via ThreadPoolExecutor (9 workers)
- Total time = slowest scraper, not sum of all
- Delays reduced to 0.5-1s (fresh GitHub IP each run)
- Enrichment capped at 15 per source for speed
- Config now accepts float delays
- Expected time: ~3-5 min instead of 12-15 min

+++ config.py
@@ -17,5 +17,5 @@ class Config:
     ADZUNA_APP_ID = os.environ.get('ADZUNA_APP_ID', '')
     ADZUNA_APP_KEY = os.environ.get('ADZUNA_APP_KEY', '')
     TIMEZONE = 'America/Toronto'
-    SCRAPE_DELAY_MIN = int(os.environ.get('SCRAPE_DELAY_MIN', 2))
-    SCRAPE_DELAY_MAX = int(os.environ.get('SCRAPE_DELAY_MAX', 5))
+    SCRAPE_DELAY_MIN = float(os.environ.get('SCRAPE_DELAY_MIN', 2))
+    SCRAPE_DELAY_MAX = float(os.environ.get('SCRAPE_DELAY_MAX', 5))
